The POC now resembles what it is supposed to be.  The gui has been changed to actually do a performance test.  There are now input fields for Number of Entries and Size of entries. A Write Test button causes the ejb in the back end to write entries to the cache and time the write operations.  The results are written back to the screen in the message field. Likewise a Read Test botton does the same thing with read operations.
The reponses are strung together on one line.  Given time, I will change this on monday so the
responses are written in rows similar to the previous swing based version of the test.

At this point, I have been running two node grids.  This has been working for a small number
of entries.  When the number of entries increase, however, I am getting state change timeout
errors.  This might be fixable via configuration or network changes.  I will take a closer look
at this on Monday.

Randy